refactor: remove unnecessary features and fix language change
- Remove Habits feature (tracker, streaks, logging)
- Remove Mood Tracker feature (calendar heatmap, insights)
- Remove Leaderboard (mock with fake opponents)
- Remove Academic Section (hardcoded subjects)
- Fix language change: disable RTL layout flip, only fonts change
- Remove slow recreate() on language switch
- Update theme to force LTR layout direction
- Preserve database migrations for backward compatibility

+++ app/src/main/java/com/benyaminrasouli/phoenixprotocol/core/data/db/dao/FocusSessionDao.kt
@@ -23,7 +23,7 @@ interface FocusSessionDao {
     suspend fun getSessionById(sessionId: Long): FocusSession?
 
     @Query("SELECT COALESCE(SUM(durationSeconds), 0) FROM focus_sessions WHERE completed = 1")
-    suspend fun getTotalFocusTimeSeconds(): Int
+    suspend fun getTotalFocusTimeSeconds(): Long
 
     @Query("SELECT COUNT(*) FROM focus_sessions WHERE completed = 1")
     suspend fun getCompletedSessionCount(): Int
